luong tao order

+++ src/main/java/com/vt/cms/model/entity/Payment.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 @Data
 public class Payment {
     private String paymanrtid;
-    private Long OrderId;
+    private Integer OrderId;
     private String status;
     private  String trancactioncode;
     private BigDecimal amount;
